Revert "In MODULARIZE mode avoid modifying the incoming moduleArg. NFC (#21775)" (#21994)
This change effectively reverts 4dac6d6d76.

It turns out that there is at least one use case where accessing the
partially constructed module prior to the promise resolution is
necessary.  Specifically when using `--preload-files` the file packager
output expects to be able to use the `Module.FS` functions before the
module is loaded.  This PR adds a test for this scenario.