Re-enable extern-pre-js and extern-post-js under pthreads
In #21701 the running of extern-pre/post-js in pthreads was disabled.
The reason was that we had a couple of tests that seemed to be assuming
this.  However, it turns out that there are important use cases that
depend on this code running in threads.

Instead, fix the two test cases by adding an extra condition.

Fixes: #22012