Re-enable XHR as fallback for fetch on webview
In #22015 we switched `readAsync` to be based on fetch rather than
XHR.  However that broke the assumption in `getBinaryenPromise` that was
assuming `readAsync` was XHR-based.

This change reinstates the XHR fallback in the case of webviews where
loading from `file://` URLs is not only possible but (apparently)
common.

Sadly we don't have any webview tests so I have no way to actually test
this within our current frameworks.