Improve comment preservation with in acorn-optimizer. NFC
Prior to this change we would drop any comment that was more than 20
characters from JS symbol, based on the end of the commend and that
start of the symbol

With this change we map all comments to the start of the next symbol.
This means that we can have many comments in row that all map the same
location and we decide to either keep all of them or discard them.