Use C/C++ bool type to implement EM_BOOL (#22157)
This reduces the size of several structs and can result in code size
savings in some cases. The reason
the code size savings don't show up in trivial examples is (I believe)
because this change also increases
the use of HEAP8 (where previously some examples only depended on
HEAP32).

This change is split of from a larger change I have planned to remove
the use of EM_BOOL completely: #22155.