Prototype module splitting analysis using relocations
This is a prototype script implementing analysis for module splitting using the
symbol table and relocations.
It builds a dependence graph that includes all function and data symbols
(as defined in the symbol table when linking with the --emit-relocs flag)

Then, given a list of functions to use as the entry points to a module,
it finds the list of functions that are reachable from the module entry
points but not reachable from main (except via paths that pass through
the module entry points; this alllows for the case where module
entry points are themselves reachable from main).

It has an alternate mode where an 'anchor' function can be specified
to exclude functions reachabe from the anchor rather than main; and
a utility to print all paths that reach a specified function without
going through the module entry points (for understanding why a
particular function is being retained in the main module).

All of this is still an experimental prototype, but it demonstrates
concept, and with a little massaging of the output, the list of
functions could be fed directly to the current version of wasm-split.
(This has not been tested though).