Save intermediate object files when `--save-temps` is used. NFC (#22239)
clang takes care of saving the `.s`, `.i` and `.bc` files but its up to
the compiler driver (in this case emcc) to save the temporary object
files.

Interestingly if two source files have the same basename then `clang
--save-temps` doesn't work since the object files will clobber each
other in the current working directory. This is probably a bug in clang
so I didn't recreate it here.