Improve SCons support (#18945)
Made it so that `emscons` also provides the `EMSCRIPTEN_ROOT`
environment variable and made the SCons tool provide the environment
variables that `emmake` (and others) does.

I was in the spirit of #18279 and started making the SCons tool search
for the root directory by scanning the `PATH` instead of relying on that
variable, but then I realised that it would be much simpler for
`emscons` to provide it.

I also noted that, unlike `emmake` and `emconfigure`, the `emscons`
command would not set some environment variables — notably to me, the
`pkg-config`-related ones.

Because the default SCons behaviour is to not propagate environment
variables to child processes, it made little sense to add them in
`emscons.py`. Instead I made the SCons tool directly import the relevant
Python package from the root directory and get only the relevant
environment variables from there, so that it would not touch any other.

For that I had to create a different function, because
`get_building_env()` returns the inherited environment variables as
well.

For the tests I made the SCons script aware of what environment
variables to expect by passing to it an argument containing them in JSON
form. Is that too hacky?