Fix partial write case when using _fd_write/doWritev/writev
doWritev iterates through the iovs structure writing the data
for each entry. The bug fixed here is that if a write does not
write all data, the loop continues. If the subsequent write also
accepts data, it has accepted data out of order.

Without the fix to doWritev this new test outputs:

```
TO DEVICE: "A"
TO DEVICE: "X"
TO DEVICE: "C"
TO DEVICE: "X"
TO DEVICE: "Y"
TO DEVICE: "Z"
TO DEVICE: "a"
TO DEVICE: "\n"
TO DEVICE: "c"
TO DEVICE: "\n"
done
```

Fixes #22258