Remove old memcpy test. NFC (#22295)
This test was added back in 2011 (00d573c696) back when we implemented
things like memcpy directly in asm.js. I'm not sure it has any value
today and its pretty hard to see what its trying to verify. It also
check its output using an md5sum which means that it if it ever failed
it would be near impossible to see why.

This change was motivated by a desire to the output_nicerizer argument
to the test runner, of which this is the only real user.