[webview] prevent potential double resolution in XMLHttpRequest handling (#22404)
Although the JavaScript engine can handle this situation, attempting to
handle both `resolve` and `reject` for the same `Promise` in the code
logic indicates that the code structure is not sufficiently rigorous and
may have design flaws. This can obscure the intent of the code and
potentially confuse other developers.