Enable SINGLE_FILE + PTHREADS html output (#22547)
In SINGLE_FILE mode, the JS is inlined directly into the generated HTML.
However for pthreads builds we need to be able to create workers that
use the same inlined JS. To enable this we encode the JS file into a
data URL that can be used for the script `src` attribute. The pthread
worker can then use this `src`.

Prior to this change using SINGLE_FILE + PTHREADS + html output would
simply fail.