Revert "[tsgen] Handle different get/set types for value_objects too." (#22569)
Reverts emscripten-core/emscripten#22439

This was working okay for when the user gets a value_object in TS, but
when setting a value_object with an object literal this code does not
type check correctly. There's an [upstream issues about this](51229). I
think instead of generating getter/setters we can instead have two type
definitions for each value_object. One for reading from TS and one for
write from TS. Until I get around to trying that I think we can revert
this in the meantime.