[file-packager] Use `fetch()` for loading file packages. (#22016)
This PR switches AJAX requests for file packages to be loaded with
fetch() rather than XmlHttpRequest(). This allows service workers to
make AJAX requests, since XmlHttpRequest is not available in service
workers.

`Module['setStatus']` is now called with more granularity during package
downloads.

Previously, --embed-file was necessary to load file packages, which
embeds the files into the binary and precludes caching and re-use of the
package.

Partially fixes
https://github.com/emscripten-core/emscripten/issues/22003, only applies
to file packages.