Move main loops handling code to library_eventloop.js. NFC
This means that programs that use the main event loops don't need to
depend on the `Browser` object.

We already supported and test the main loops handling code on node, but
programs that only use the event loop without any other `Browser`
functions no longer need to pull in the whole `Browser` object.