Fix combining of test parametization. NFC
As far as I can tell the only test that was relying on this features
is `browser.test_sdl_image_prepare` which uses `@also_with_proxied` in
addition to `@also_with_wasmfs` and it was getting the names and values
backwards.  I confirmed manually that all 4 test combination now do the
right thing.