Don't listen to 0.0.0.0 by default for security reasons. (#22077)
A developer might start emrun without thinking about the --hostname
setting. So he might expose sensible data on the webserver to the LAN or
to the Internet. It might even happen, that a vulnerable Emsdk version
becomes publicly reachable.

Instead use localhost (127.0.0.1) by default, which is usually
sufficient for development.