Let leakWarning carry the string representation, not the stack itself (#22488)
At least with recent Chrome, I ran into a situation where an object of
mine apparently didn't get garbage collected because it was retained via
such a leakWarning stack trace. (I'm not sure about the internal
workings of that, though; but the object started to get garbage
collected once I did this change.)

I have no idea if there are any other up-/downsides to carrying the
stack vs. just the string here. And my explanation of what exactly it
fixed for me is probably a bit vague. But wanted to bring it up
nevertheless---maybe it is considered a good change anyway.