[test] Re-instate ability to delay reference image collection. NFC (#22672)
In #22665 I removed the manually_trigger_reftest. However this resulted
in racey behaviour in some tests. It turns out that running `doReftest`
during `postRun` was working, but only because `doReftest` itself was
doing async work, thus delaying the actual capture of the canvas bits.

This change re-instates the behaviour but without the need to test
setting in python.