Allow `SHARED_MEMORY` flag to be used independently.
Currently, there are a couple of restrictions on the `SHARED_MEMORY` flag:
* Attempting to use the `SHARED_MEMORY` flag without the `PTHREADS` or
  `WASM_WORKERS` flags simply fails, because we don't have variations of
  the system libraries that have this flag standalone. Instead, emcc
  attempts to use the single-threaded system libraries, which are not
  compatible with the user's build because they don't have the correct
  wasm feature flags (i.e. -matomics and -mbulk-memory). This PR adds
  `SHARED_MEMORY`-only variations of the system libraries, which allows
  users to build with a standalone `SHARED_MEMORY` flag without error.
* Currently, `SHARED_MEMORY` also implies `IMPORTED_MEMORY`. However,
  `IMPORTED_MEMORY` is only important for builds that use the `PTHREADS`
  or `WASM_WORKERS` flags, since they actually do multi-threading and
  need to read and write to the same memory across workers. This PR
  changes this so that if a standalone `SHARED_MEMORY` flag is specified,
  it can still avoid importing the memory.

The use-case for enabling the `SHARED_MEMORY` flag without actually having
an actual threading implementation is to be compatible with other modules
which may import that memory whose import declaration specifies the memory
as shared. This is an important use-case for Flutter, as we want to create
two variations of the emscripten-generated module (one single-threaded and
one multi-threaded, depending on whether the environment supports it) while
keeping the module that imports the emscripten-generated module the same
(which means keeping the memory import declaration as shared).