Fix audio worklet + memory64 (#22731)
There are a two different fixes here, neither of which are needed except
in the nightly version of chrome.

1. In the `instantiateWasm` override used in audio worklets return the
result of `receiveInstance` rather then directly returning the wasm
exports. This means that the resulting object has the needed wrappers
applied. This was causing an exception even on older version of chrome,
but the tests still passed regardless. With newer versions of chrome the
exception seems to cause the worklet creation to fail. (see #18711)
2. Using BigInt when indexing the wasm table. I would have prefered to
use `toIndexType` here but that doesn't work in `audio_worklet.js`.