allow more features for maybe-wasm2js and wasm-split
This is in preparation for alloing bulk memory and nontrapping-fp by default.
Maybe-wasm2js just allows all features for simplicity (although if a feature
unsupported by wasm2js is used, the test may fail anyway).
The wasm-split test fails when reference types are enabled; I haven't
investigated that yet, but for now we just enable the 2 we are interested in.