Remove automatic fallback to $HOME/.emscripten_cache. NFC
Putting the cache in the $HOME directory is not a great idea since its
a location that is shared by any/all versions of emscripten.  If anyone
really wants to use that location they can do so via the config file
setting `CACHE` or via the environment variable `EMCC_CACHE`.