Remove expectation requiring no non-debug custom sections in separate-dwarf files
This test is now failing after https://github.com/WebAssembly/binaryen/pull/7043
because the target_features section is in the output binary and gets copied
over to the separate-dwarf file (this is because the separate-dwarf file is
just a copy of the original file).

An alternative to this would be to strip all non-debug custom sections
from the side file, although there could be sections that emscripten
doesn't know about that it would be desirable to have.