Reduce usage of legacy `event.keyCode`. NFC
This change replaced the usage of the legacy `keyCode` attribute
with the preferred `key` attribute, which also has the advantage of
removing some hardcoded constant numbers.

In order to test this change I fixed the `test_glfw_get_key_stuck`
test and added `test_sdl_key_test` to the interactive tests.

In doing so I noticed and fixed a crash bug that was introduced in #22874.