Use quoted access for `Module['ctx']`. NFC
This was the last property explicitly listed in the closure type
annotate for the `Module` object.  This means that it was previously
being preserved by closure via this mechanism, but now we preserve it
via the normal quoting mechanism.

I've been looking at removing `ctx` and `canvas` from the `Module`
object, but this change is just about consistency.

See #22876