Log SDL keyboard events as part of RUNTIME_DEBUG. (#22918)
Also, test_sdl_key_test.c to test_interactive.py. This test was not
being run anywhere else and serves as simple way to test/debug keyboard
events.

Also, only set the `RUNTIME_DEBUG` default if the user does not
explicitly set it themselves. This allows for `-sSOCKETS_DEBUG
-sRUNTIM_DEBUG=0` to opt out of the `RUNTIME_DEBUG` channel.

Finally, disable the `RUNTIME_DEBUG` traces that occur during turns of
the event loop unless `RUNTIME_DEBUG` is set to 2. This avoids flooding
the console will messages in applications that are using the event loop.