[GLFW] Restore css scaling behavior (removed in 3.1.51) (#22900)
- the changes in 3.1.51 to add HiDPI support removed a feature that is
not documented nor a feature of GLFW
- raylib wants it back so this patch restores it

Some points:

* I added a function `isCSSScalingEnabled()` with documentation to make
it clear what is happening
* Due to the fact that HiDPI / 4k support overrides CSS, CSS scaling is
disabled in that mode. There would be a conflict otherwise
* Since this feature is not really standard, I introduced a way to
disable it (since it was implemented prior to 3.1.51, I felt it was
better to have it on by default with a way to disable it rather than the
other way around).

I have tested the changes as best as I could with ImGui and raylib. I
tested a few examples with raylib and it seems to be doing what they
want (the canvas occupies the fullscreen and the mouse coordinates are
correct and resizing the window follows appriopriately)

FIxes: #22847