Avoid try/catch in findEventTarget. NFC
Rather than using try/catch to detect when `document` is missing
we explicitly check for it.  This is the same pattern used in the other
version of `findEventTarget` on line 336.

This avoids casting a wide net and hiding errors in the rest of the code
in this function.  Its also more consistent with the other version of
`findEventTarget`.

One change here is that the fallback for when `querySelector` fails is
now the same as the fallback for when `document` is missing.  In this
case we return `undefined`.  Returning a string from findEventTarget
is never correct.  Its always supposed to return a DOM element AFAICT.