Avoid try/catch in findEventTarget. NFC (#22984)
Rather than using try/catch to detect when `document` is missing we
explicitly check for it. This is the same pattern used in the other
version of `findEventTarget` on line 336.

This avoids casting a wide net and hiding errors in the rest of the code
in this function. Its also more consistent with the other version of
`findEventTarget`.