Remove cleanup code from stdlib tests. NFC
When running tests in emscripten there is no requirement to clean up
after yourself.  The test framework automatically gives each test
a clean directory to work in.

In fact, cleaning up after yourself like this makes debugging harder
because in the failure case you can not longer see the state of the
filesystem when test fails.

In addition, it meant that all these tests were also relying on and
testing `atexit` and `signal` handling, making the tests more complex
that less precise.

As part of this change I also make sure all these tests could be
compiled outside of emscripten (on my desktop linux machine).