Remove basic_string<unsigned_char> from embind
Only `char`, `wchar`, `char8`, `char16`, and `char32` are valid
specialization for `std::basic_string`:
https://en.cppreference.com/w/cpp/string/basic_string

But libc++ had a base template for `basic_string` that allowed any type
to be passed for a long time. It looks there have been several attempts
to remove this after which they restored it due to complaints, in
chronological order:
https://github.com/llvm/llvm-project/commit/aeecef08c385b1e4955155dd649a2d3724463849
https://github.com/llvm/llvm-project/commit/08a0faf4cd32bce6c51027ea9b5ec351747995b4
https://github.com/llvm/llvm-project/commit/e30a148b098d462d0267c479cd9e4783363c2761
https://github.com/llvm/llvm-project/pull/66153
https://github.com/llvm/llvm-project/pull/72694

The last one, https://github.com/llvm/llvm-project/pull/72694,
eventually removed it. So `std::basic_string<unsigned_char>` is not
allowed anymore. This removes all uses of
`std::basic_string<unsigned_char>` from embind.

This needs to be done to update libc++ to LLVM 19 (#22994). I'm
uploading this as a separate PR because this removes a functionality
from embind.