Fix other.test_embind_no_duplicate_symbols
After https://github.com/llvm/llvm-project/pull/80282,
`formatter_integral.h` includes not `locale` but `_locale`
https://github.com/llvm/llvm-project/blame/aadaa00de76ed0c4987b97450dd638f63a385bed/libcxx/include/__format/formatter_integral.h#L35
and after several more include changes this causes
https://github.com/llvm/llvm-project/blob/aadaa00de76ed0c4987b97450dd638f63a385bed/libcxx/include/__thread/support/pthread.h
not to be included when
https://github.com/emscripten-core/emscripten/blob/main/system/include/emscripten/val.h
is included. `pthread.h` is where `pthread_t` is defined, which is used
here:
https://github.com/emscripten-core/emscripten/blob/fb14f6d9fc2b00fea36af0f5ca4e996f61520b8f/system/include/emscripten/val.h#L676

Until LLVM 18, the contents of
https://github.com/llvm/llvm-project/blob/aadaa00de76ed0c4987b97450dd638f63a385bed/libcxx/include/__thread/support/pthread.h
was in
https://github.com/llvm/llvm-project/blob/3b5b5c1ec4a3095ab096dd780e84d7ab81f3d7ff/libcxx/include/__threading_support,
which was split off to several files in
https://github.com/llvm/llvm-project/pull/79654. This file was
(transitively) included from `locale`.

Anyway, this commit just adds `#include <pthread.h>` directly to
`val.h`.