Only run other.test_minimal_runtime_code_size_hello_embind
Currently `other.test_minimal_runtime_code_size_hello_embind` fails on
CircleCI but not on my local machine. The expected size of `a.wasm` is
greater on CircleCI but even their numbers are all different.

The expected total size is 17651, which is also what my local machine's
result.
But in CI, it is 17655 in test-other, 17659 in test-mac-arm64, and 17667
in test-windows.

I tried to log in to ssh in test-other CI and ran
`other.test_minimal_runtime_code_size_hello_embind` manually in that
environment. But the result's size is 17651, the same as the expected
size, which is baffling.

This commit is trying to make the CI only run
`other.test_minimal_runtime_code_size_hello_embind` so that I can access
the contents of `out/test` directory right after that execution. This is
only for testing and will be reverted later.