Remove `runDependencies` handling code from  `emscripten_async_load_script`
This handling of `runDependencies` was included when this code was
first added in 1fc6762e3.  While there may have been reason for this
back then I can't see how or why this is useful today.  We don't have
any internal usage of `emscripten_async_load_script` so there should
be no way to call it before the program is running.