JS optimizer: Fix scoping issues with non-defined functions (#23159)
Other than simple defined functions (function foo() {}), JS also has
arrow functions and object methods. We need to be aware of those
in metadce.

This fixes #22968 which was an issue with an object method that
enclosed a function of the same name as another toplevel function.
To fix this, this PR makes us aware of such scopes, and we do not
optimize functions in them. That is, we only ever optimize defined
functions at the toplevel scope, and anything enclosed is considered
fixed.