Don't inject assertions onto the ready promise
This is a debug feature that means users who did not await the promise
and tried to access stuff directly on it would see a nice error message.

However with the advent of more usage of async/await, for example in
\#23157 it is not always the ready promise that gets returned to the
user but a language-generated promise from an `await` call.

We could try to jump through more hoops to continue to support this
debug feature, but I don't think added complexity in the toolchain
or the generated code would be worth it.

Split out from #23157.